requirements.txt
pip freeze
altair==5.5.0
anyio==4.11.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
bcc==0.18.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==6.3.0
blinker==1.9.0
Bottleneck==1.6.0
Brlapi==0.8.3
Brotli==1.0.9
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
chardet==4.0.0
charset-normalizer==3.4.4
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
comm==0.2.3
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.8.17
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
et-xmlfile==1.0.1
exceptiongroup==1.3.0
executing==2.2.1
fastjsonschema==2.21.2
fonttools==4.29.1
fqdn==1.5.1
fs==2.4.12
gast==0.5.2
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
hidpidaemon==18.4.6
html5lib==1.1
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
idna==3.11
importlib-metadata==4.6.4
iniconfig==1.1.1
ipykernel==7.1.0
ipython==8.37.0
ipython_genutils==0.2.0
isoduration==20.11.0
jdcal==1.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.0.3
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
kernelstub==3.1.4
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
lark==1.3.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llvmlite==0.38.0
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.2.1
mistune==3.1.4
more-itertools==8.10.0
mpmath==0.0.0
narwhals==2.10.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
notebook==7.4.7
notebook_shim==0.2.4
numba==0.55.1
numexpr==2.14.1
numpy==2.2.6
oauthlib==3.2.0
odfpy==1.4.2
olefile==0.46
openpyxl==3.0.9
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
Pillow==9.0.1
platformdirs==4.5.0
plotly==6.4.0
pluggy==0.13.0
ply==3.11
pop-transition==1.1.2
prometheus_client==0.23.1
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.2
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.10.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.20.1
pycparser==2.23
pycups==2.0.1
pydbus==0.6.0
pydeck==0.9.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==6.2.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-gnupg==0.4.8
python-json-logger==4.0.0
python-xlib==0.29
pythran==0.10.0
pytz==2025.2
pyxdg==0.27
PyYAML==5.4.1
pyzmq==27.1.0
referencing==0.37.0
repolib==2.2.1
repoman==1.4.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.28.0
rsa==4.9.1
scipy==1.8.0
SecretStorage==3.3.1
Send2Trash==1.8.3
sessioninstaller==0.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.3.1
stack-data==0.6.3
streamlit==1.51.0
sympy==1.9
systemd-python==234
tables==3.7.0
tenacity==9.1.2
terminado==0.18.1
tinycss2==1.4.0
toml==0.10.2
tomli==2.3.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unicodedata2==14.0.0
uri-template==1.3.0
urllib3==2.5.0
wadllib==1.3.6
watchdog==6.0.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
xdg==5
xkit==0.0.0
xlwt==1.3.0
zipp==1.0.0
